refactor: Extract search utils from Sections
